feature(devops): Add dist directory of front-end to statics, and serve the Angular app.
When the Angular app is built, its deployment is automatic through the collectstatic management command.
The root URL serve a Django template that load the Angular app, using Whitenose.
Because the collecstatic management command is executed on Heroku on each push, the deployment of the back and front ends are automatic and simultaneous.